Refactor API structure by creating separate routers for URLs and QR
